projects/locals-only-warp/pyproject.toml
[project]
name = "locals-only-warp"
version = "0.1.0"
description = "Local business and activity recommendations app"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask>=2.3.3",
    "langchain>=0.0.27",
    "langchain-anthropic>=0.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "geopy>=2.3.0",
    "folium>=0.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
